Refactor init/create
