+++ src/dso/_util.py
@@ -8,7 +8,6 @@ from contextlib import contextmanager
 from functools import cache
 from os import environ
 from pathlib import Path
-from typing import TYPE_CHECKING
 
 from rich.prompt import Confirm
 
@@ -19,12 +18,6 @@ try:
     import tomllib
 except ImportError:
     import tomli as tomllib
-
-
-if TYPE_CHECKING:
-    from importlib.resources.abc import Traversable
-
-DEFAULT_BRANCH = "master"
 
 
 def check_project_roots(paths: Sequence[Path]) -> Path:
@@ -103,72 +96,6 @@ def get_project_root(start_directory: Path) -> Path:
     else:
         # .parent, because proj_root points to the git directory
         return proj_root.parent
-
-
-def get_template_path():
-    # TODO remove
-    raise NotImplementedError
-
-
-def _copy_with_render(source: Traversable, destination: Path, params: dict):
-    """Fill all placeholders in a file with jinja2 and save file to destination"""
-    from jinja2 import StrictUndefined, Template
-
-    with source.open() as f:
-        template = Template(f.read(), undefined=StrictUndefined)
-    rendered_content = template.render(params)
-    with destination.open("w") as file:
-        file.write(rendered_content)
-        # Non-empty files should have a terminal new-line to make the pre-commit hooks happy
-        if len(rendered_content):
-            file.write("\n")
-
-
-def instantiate_template(template_path: Traversable, target_dir: Path | str, **params) -> None:
-    """Copy a template folder to a target directory, filling all placeholder values."""
-    from jinja2 import Template
-
-    target_dir = Path(target_dir)
-
-    def _traverse_template(curr_path, subdir):
-        for p in curr_path.iterdir():
-            if p.is_file():
-                name_rendered = Template(p.name).render(params)
-                # this file is used for checking in empty folders in git.
-                if name_rendered != ".gitkeeptemplate":
-                    target_file = target_dir / subdir / name_rendered
-                    if not target_file.exists():
-                        _copy_with_render(p, target_file, params)
-            else:
-                (target_dir / subdir / p.name).mkdir(exist_ok=True)
-                _traverse_template(p, subdir / p.name)
-
-    _traverse_template(template_path, Path("."))
-
-
-def instantiate_with_repo(template: Traversable, target_dir: Path | str, **params) -> None:
-    """Create a git repo in a directory and render a template inside.
-
-    Creates an initial commit.
-    """
-    from git.repo import Repo
-
-    target_dir = Path(target_dir)
-    target_dir.mkdir(exist_ok=True)
-    log.info("Created project directory.")
-
-    instantiate_template(template, target_dir, **params)
-    log.info("Created folder structure from template.")
-
-    if not (target_dir / ".git").exists():
-        Repo.init(target_dir)
-        repo = Repo(target_dir)
-        # set main as default branch
-        repo.git.checkout("-b", DEFAULT_BRANCH)
-        repo.git.symbolic_ref("HEAD", f"refs/heads/{DEFAULT_BRANCH}")
-        repo.git.add(A=True)
-        repo.index.commit("Initial commit generated by dso CLI tool.")
-        log.info("Initalized local git repo.")
 
 
 def git_list_files(dir: Path) -> list[Path]:
