+++ src/dso/templates/index.json
@@ -26,7 +26,7 @@
             "params": [
                 {
                     "name": "name",
-                    "description": "Folder name"
+                    "description": "Folder name, e.g. 'RNAseq'"
                 }
             ]
         }
